functions/data-processing-engines/dataproc-serverless-app-executor/requirements.txt
functions-framework==3.*
google-auth
google-cloud-bigquery
google-cloud-logging
grpcio
grpcio-tools
google-api-python-client
google-cloud-dataform
google-cloud-resource-manager